requirements.txt
prometheus-api-client
datadog-api-client
kubernetes
pyyaml
